feat: сделал отрисовку на шейдерах

+++ lw3/engine/src/render/Vec2f.h
@@ -2,5 +2,5 @@
 
 struct Vec2f
 {
-	double x = 0, y = 0;
+	float x = 0, y = 0;
 };
